From David Megginson:
1. Fixed a bug reported by Curt for one of the getChild methods.

2. Cleaned up some of the error reporting, and require PropertyList as
the root element name.  Also uncluttered the XML output a little
(removed type="unknown").

3. Added a new Long type (i.e. setLongValue, getLongValue, etc.), as
requested by Curt aeons ago.

4. Added aliases, so that a property's value can directly reflect the
value of another property.

5. Added a generalized inclusion facility, with parameters (panel
writers should love this one).

+++ simgear/misc/props_test.cxx
@@ -317,7 +317,7 @@ test_property_nodes ()
   cout << "Looking for all /hack[0]/bar children" << endl;
   vector<SGPropertyNode *> bar = child->getChildren("bar");
   cout << "There are " << bar.size() << " matches" << endl;
-  for (int i = 0; i < bar.size(); i++)
+  for (int i = 0; i < (int)bar.size(); i++)
     cout << bar[i]->getName() << '[' << bar[i]->getIndex() << ']' << endl;
   cout << endl;
